delete deprecated event module
