+++ src/main/java/com/trymad/litechess_monolith/chessgame/ChessPartyService.java
@@ -2,6 +2,8 @@ package com.trymad.litechess_monolith.chessgame;
 
 import java.util.List;
 
+import com.trymad.litechess_monolith.chessgame.api.dto.ChessPartyDTO;
+import com.trymad.litechess_monolith.chessgame.api.dto.CreatePartyDTO;
 import com.trymad.litechess_monolith.chessgame.api.event.GameFinishEvent;
 import com.trymad.litechess_monolith.matchmaking.api.event.GameFindedEvent;
 import com.trymad.litechess_monolith.websocket.api.event.MoveEvent;
