--- src/main/java/com/trymad/litechess_monolith/chessgame/CreatePartyDTO.java
@@ -1,5 +1,0 @@
-package com.trymad.litechess_monolith.chessgame;
-
-public record CreatePartyDTO(PlayerInfo white, PlayerInfo black) {
-	
-}
