--- src/main/java/com/trymad/litechess_monolith/chessgame/ShortChessPartyDTO.java
@@ -1,9 +1,0 @@
-package com.trymad.litechess_monolith.chessgame;
-
-public record ShortChessPartyDTO(
-	Long id,
-	PlayerInfo white,
-	PlayerInfo black,
-	ChessGameStatus status) {
-	
-}
