+++ src/main/java/com/trymad/litechess_monolith/chessgame/api/dto/CreatePartyDTO.java
@@ -1,0 +1,7 @@
+package com.trymad.litechess_monolith.chessgame.api.dto;
+
+import com.trymad.litechess_monolith.chessgame.PlayerInfo;
+
+public record CreatePartyDTO(PlayerInfo white, PlayerInfo black) {
+	
+}
