+++ src/main/java/com/trymad/litechess_monolith/chessgame/api/dto/ShortChessPartyDTO.java
@@ -1,0 +1,12 @@
+package com.trymad.litechess_monolith.chessgame.api.dto;
+
+import com.trymad.litechess_monolith.chessgame.ChessGameStatus;
+import com.trymad.litechess_monolith.chessgame.PlayerInfo;
+
+public record ShortChessPartyDTO(
+	Long id,
+	PlayerInfo white,
+	PlayerInfo black,
+	ChessGameStatus status) {
+	
+}
