+++ src/main/java/com/trymad/litechess_monolith/chessgame/internal/game/InMemoryChessPartyRepository.java
@@ -10,7 +10,7 @@ import org.springframework.stereotype.Component;
 
 import com.trymad.litechess_monolith.chessgame.ChessGameStatus;
 import com.trymad.litechess_monolith.chessgame.ChessParty;
-import com.trymad.litechess_monolith.chessgame.CreatePartyDTO;
+import com.trymad.litechess_monolith.chessgame.api.dto.CreatePartyDTO;
 
 
 @Component
