--- src/main/java/com/trymad/litechess_monolith/event/queueLeave/package-info.java
@@ -1,7 +1,0 @@
-
-@ApplicationModule(
-  type = ApplicationModule.Type.OPEN
-)
-package com.trymad.litechess_monolith.event.queueLeave;
-
-import org.springframework.modulith.ApplicationModule;
