+++ src/main/java/com/trymad/litechess_monolith/websocket/internal/GameMessageSender.java
@@ -13,7 +13,7 @@ import com.trymad.litechess_monolith.chessgame.ChessParty;
 import com.trymad.litechess_monolith.chessgame.api.event.ChessPartyCreatedEvent;
 import com.trymad.litechess_monolith.chessgame.api.event.GameFinishEvent;
 import com.trymad.litechess_monolith.chessgame.api.event.MoveAcceptedEvent;
-import com.trymad.litechess_monolith.event.queueLeave.QueueLeaveEvent;
+import com.trymad.litechess_monolith.matchmaking.api.event.QueueLeaveEvent;
 import com.trymad.litechess_monolith.shared.event.EventPublisher;
 import com.trymad.litechess_monolith.websocket.GameCreatedDTO;
 import com.trymad.litechess_monolith.websocket.MoveRequest;
